[Rename] Rename HeaderNotice's relation atoms
